requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
numpy==1.23.1
pandas==2.0.3
pony==0.7.17
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.24.0
watchfiles==0.21.0
websockets==12.0